Move response handling of lsp messages to background thread
